LPS-53829 Refactor package
