+++ modules/portal/portal-workflow-kaleo-action/src/main/java/com/liferay/portal/workflow/kaleo/action/executor/internal/DRLActionExecutor.java
@@ -12,21 +12,24 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.action.executor.script;
+package com.liferay.portal.workflow.kaleo.action.executor.internal;
 
-import com.liferay.portal.kernel.scripting.ScriptingUtil;
+import com.liferay.portal.kernel.bi.rules.Fact;
+import com.liferay.portal.kernel.bi.rules.Query;
+import com.liferay.portal.kernel.bi.rules.RulesEngineUtil;
+import com.liferay.portal.kernel.bi.rules.RulesResourceRetriever;
+import com.liferay.portal.kernel.resource.StringResourceRetriever;
+import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutor;
+import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutorException;
 import com.liferay.portal.workflow.kaleo.model.KaleoAction;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
-import com.liferay.portal.workflow.kaleo.runtime.action.ActionExecutor;
-import com.liferay.portal.workflow.kaleo.runtime.action.ActionExecutorException;
-import com.liferay.portal.workflow.kaleo.runtime.util.ScriptingContextBuilderUtil;
+import com.liferay.portal.workflow.kaleo.runtime.util.RulesContextBuilder;
 import com.liferay.portal.workflow.kaleo.util.WorkflowContextUtil;
 
 import java.io.Serializable;
 
-import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -35,19 +38,9 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	immediate = true,
-	property = {
-		"com.liferay.portal.workflow.kaleo.action.script.language=beanshell",
-		"com.liferay.portal.workflow.kaleo.action.script.language=groovy",
-		"com.liferay.portal.workflow.kaleo.action.script.language=javascript",
-		"com.liferay.portal.workflow.kaleo.action.script.language=python",
-		"com.liferay.portal.workflow.kaleo.action.script.language=ruby"
-	}
+	property = {"com.liferay.portal.workflow.kaleo.action.script.language=drl"}
 )
-public class ScriptActionExecutor implements ActionExecutor {
-
-	public ScriptActionExecutor() {
-		_outputObjects.add(WorkflowContextUtil.WORKFLOW_CONTEXT_NAME);
-	}
+public class DRLActionExecutor implements ActionExecutor {
 
 	@Override
 	public void execute(
@@ -63,21 +56,21 @@ public class ScriptActionExecutor implements ActionExecutor {
 		}
 	}
 
-	public void setOutputObjects(Set<String> outputObjects) {
-		_outputObjects.addAll(outputObjects);
-	}
-
 	protected void doExecute(
 			KaleoAction kaleoAction, ExecutionContext executionContext,
 			ClassLoader... classLoaders)
 		throws Exception {
 
-		Map<String, Object> inputObjects =
-			ScriptingContextBuilderUtil.buildScriptingContext(executionContext);
+		List<Fact<?>> facts = RulesContextBuilder.buildRulesContext(
+			executionContext);
 
-		Map<String, Object> results = ScriptingUtil.eval(
-			null, inputObjects, _outputObjects, kaleoAction.getScriptLanguage(),
-			kaleoAction.getScript(), classLoaders);
+		RulesResourceRetriever rulesResourceRetriever =
+			new RulesResourceRetriever(
+				new StringResourceRetriever(kaleoAction.getScript()));
+
+		Map<String, ?> results = RulesEngineUtil.execute(
+			rulesResourceRetriever, facts, Query.createStandardQuery(),
+			classLoaders);
 
 		Map<String, Serializable> resultsWorkflowContext =
 			(Map<String, Serializable>)results.get(
@@ -86,7 +79,5 @@ public class ScriptActionExecutor implements ActionExecutor {
 		WorkflowContextUtil.mergeWorkflowContexts(
 			executionContext, resultsWorkflowContext);
 	}
-
-	private final Set<String> _outputObjects = new HashSet<>();
 
 }
