+++ modules/portal/portal-workflow-kaleo-action/src/main/java/com/liferay/portal/workflow/kaleo/action/executor/internal/ScriptActionExecutor.java
@@ -12,24 +12,21 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.action.executor.script;
+package com.liferay.portal.workflow.kaleo.action.executor.internal;
 
-import com.liferay.portal.kernel.bi.rules.Fact;
-import com.liferay.portal.kernel.bi.rules.Query;
-import com.liferay.portal.kernel.bi.rules.RulesEngineUtil;
-import com.liferay.portal.kernel.bi.rules.RulesResourceRetriever;
-import com.liferay.portal.kernel.resource.StringResourceRetriever;
+import com.liferay.portal.kernel.scripting.ScriptingUtil;
+import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutor;
+import com.liferay.portal.workflow.kaleo.action.executor.ActionExecutorException;
 import com.liferay.portal.workflow.kaleo.model.KaleoAction;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
-import com.liferay.portal.workflow.kaleo.runtime.action.ActionExecutor;
-import com.liferay.portal.workflow.kaleo.runtime.action.ActionExecutorException;
-import com.liferay.portal.workflow.kaleo.runtime.util.RulesContextBuilder;
+import com.liferay.portal.workflow.kaleo.runtime.util.ScriptingContextBuilderUtil;
 import com.liferay.portal.workflow.kaleo.util.WorkflowContextUtil;
 
 import java.io.Serializable;
 
-import java.util.List;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -38,9 +35,19 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	immediate = true,
-	property = {"com.liferay.portal.workflow.kaleo.action.script.language=drl"}
+	property = {
+		"com.liferay.portal.workflow.kaleo.action.script.language=beanshell",
+		"com.liferay.portal.workflow.kaleo.action.script.language=groovy",
+		"com.liferay.portal.workflow.kaleo.action.script.language=javascript",
+		"com.liferay.portal.workflow.kaleo.action.script.language=python",
+		"com.liferay.portal.workflow.kaleo.action.script.language=ruby"
+	}
 )
-public class DRLActionExecutor implements ActionExecutor {
+public class ScriptActionExecutor implements ActionExecutor {
+
+	public ScriptActionExecutor() {
+		_outputObjects.add(WorkflowContextUtil.WORKFLOW_CONTEXT_NAME);
+	}
 
 	@Override
 	public void execute(
@@ -56,21 +63,21 @@ public class DRLActionExecutor implements ActionExecutor {
 		}
 	}
 
+	public void setOutputObjects(Set<String> outputObjects) {
+		_outputObjects.addAll(outputObjects);
+	}
+
 	protected void doExecute(
 			KaleoAction kaleoAction, ExecutionContext executionContext,
 			ClassLoader... classLoaders)
 		throws Exception {
 
-		List<Fact<?>> facts = RulesContextBuilder.buildRulesContext(
-			executionContext);
+		Map<String, Object> inputObjects =
+			ScriptingContextBuilderUtil.buildScriptingContext(executionContext);
 
-		RulesResourceRetriever rulesResourceRetriever =
-			new RulesResourceRetriever(
-				new StringResourceRetriever(kaleoAction.getScript()));
-
-		Map<String, ?> results = RulesEngineUtil.execute(
-			rulesResourceRetriever, facts, Query.createStandardQuery(),
-			classLoaders);
+		Map<String, Object> results = ScriptingUtil.eval(
+			null, inputObjects, _outputObjects, kaleoAction.getScriptLanguage(),
+			kaleoAction.getScript(), classLoaders);
 
 		Map<String, Serializable> resultsWorkflowContext =
 			(Map<String, Serializable>)results.get(
@@ -79,5 +86,7 @@ public class DRLActionExecutor implements ActionExecutor {
 		WorkflowContextUtil.mergeWorkflowContexts(
 			executionContext, resultsWorkflowContext);
 	}
+
+	private final Set<String> _outputObjects = new HashSet<>();
 
 }
